centro_emprendedores_app/backend/requirements.txt
# Centro Mundial de Emprendedores - Backend
# Dependencias específicas del backend

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Configuración
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Desarrollo
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging
loguru==0.7.2

# HTTP Client
httpx==0.25.2
requests==2.31.0
